+++ src/geo-commons/services/geojson.parser.ts
@@ -33,9 +33,58 @@ export abstract class AbstractGeoJsonParser<T extends LatLngBase> extends Abstra
         return elements;
     }
 
+    public createTrack(name: string, type: string, segments: [T[]], defaultPosition: T): string {
+        const gpxSegments: string[] = [];
+        for (const segment of segments) {
+            gpxSegments.push(
+                this.createJsonTrackSegment(segment, defaultPosition));
+        }
+
+        return this.createJsonTrack(name, type, gpxSegments);
+    }
+
+    public createRoute(name: string, type: string, points: T[], defaultPosition: T): string {
+        return this.createJsonRoute(name, type, points, defaultPosition)
+    }
+
+    public createJsonTrack(name: string, type: string, segments: string[]): string {
+        return this.createJson(name, type, segments);
+    }
+
+    public createJsonTrackSegment(points: T[], defaultPosition: T): string {
+        return this.createJsonPointSegment(points);
+    }
+
+    public createJsonRoute(name: string, type: string, points: T[], defaultPosition: T): string {
+        return this.createJson(name, type, [this.createJsonPointSegment(points)]);
+    }
+
+    protected createJson(name: string, type: string, pointSegments: string[]): string {
+        return '{ "track": {"' +
+            '"tName":""' + name + '"",' +
+            '"type":""' + type + '"",' +
+            '"header":["lat","lon","ele"],' +
+            '"records":[' +
+            pointSegments.join(',') +
+            ']' +
+            '}}';
+    }
+
+    protected createJsonPointSegment(points: T[]): string {
+        return points.map(point => this.createJsonPoint(point)).join(',')
+    }
+
+    protected createJsonPoint(point: T): string {
+        const alt = point['alt']
+            ? point['alt']
+            : undefined;
+
+        return '[' + [point.lat, point.lng, alt].join(',') + ']';
+    }
+
     protected parseJsonObj(obj, options): GeoElementBase<T>[] {
         let j;
-        const coords = [];
+        const coords: T[] = [];
 
         for (j = 0; j < obj['track']['records'].length; j++) {
             const record = obj['track']['records'][j];